+++ src/views/uBikeTable/uBikeTable.vue
@@ -1,6 +1,10 @@
 <script setup>
 import { ref, computed, watch } from 'vue';
 import "bootstrap/dist/css/bootstrap.css";
+import search from './components/search.vue';
+import uBikeTable from './components/uBikeTable.vue';
+import pagination from './components/pagination.vue';
+
 
 // 修改這份 YouBike 即時資訊表：
 // 1. 將搜尋的部分拆出來變成子元件 `uBikeTable/components/search.vue`
@@ -29,20 +33,18 @@ const currentPage = ref(1);
 const COUNT_OF_PAGE = 20;
 // 最多顯示幾頁
 const PAGINATION_MAX = 10;
+
 fetch('https://tcgbusfs.blob.core.windows.net/dotapp/youbike/v2/youbike_immediate.json')
   .then(res => res.text())
   .then(data => {
     uBikeStops.value = JSON.parse(data);
   });
-const timeFormat = (val) => {
-  // 時間格式
-  const pattern = /(\d{4})(\d{2})(\d{2})(\d{2})(\d{2})(\d{2})/;
-  return val.replace(pattern, '$1/$2/$3 $4:$5:$6');
-};
+
 // 監聽搜尋文字，若有變動則將頁碼歸 1
 watch(searchText, () => {
   currentPage.value = 1;
 });
+
 // 篩選後的站點資料
 const filtedUbikeStops = computed(() => {
   return uBikeStops.value.length === 0
@@ -69,135 +71,65 @@ const slicedUbikeStops = computed(() => {
 const totalPageCount = computed(() => {
   return Math.ceil(filtedUbikeStops.value.length / COUNT_OF_PAGE);
 });
-// 分頁的尾端
-const pagerEnd = computed(() => {
-  return totalPageCount.value <= PAGINATION_MAX
-    ? totalPageCount.value
-    : PAGINATION_MAX;
-});
-// 分頁的位移，用來確保目前的頁碼在中間
-const pagerAddAmount = computed(() => {
-  const tmp =
-    totalPageCount.value <= PAGINATION_MAX
-      ? 0
-      : currentPage.value + 4 - pagerEnd.value;
-  return tmp <= 0
-    ? 0
-    : totalPageCount.value - (PAGINATION_MAX + tmp) < 0
-      ? totalPageCount.value - PAGINATION_MAX
-      : tmp;
-});
-// 換頁
-const setPage = page => {
-  if (page < 1 || page > totalPageCount.value) {
-    return;
-  }
-  currentPage.value = page;
-};
-// 指定排序
-const setSort = sortType => {
-  if (sortType === currentSort.value) {
-    isSortDesc.value = !isSortDesc.value;
-  } else {
-    currentSort.value = sortType;
-    isSortDesc.value = false;
-  }
-};
-// 關鍵字 Highlight
-const keywordsHighlight = (text, keyword) => {
-  const reg = new RegExp(keyword, 'gi');
-  return text.replace(reg, `<span style="color: red;">${keyword}</span>`);
-};
+
+
+
+
+
+//search元件註冊的事件
+const updateSearchText = (val) => {
+  searchText.value = val;
+}
+
+//uBikeTable元件註冊的事件
+const updateCurrentSort = (val) => {
+  currentSort.value = val;
+}
+const updateIsSortDesc = (val) => {
+  isSortDesc.value = val;
+}
+
+//pagination元件註冊的事件
+const updateCurrentPage = (val) => {
+  currentPage.value = val;
+}
 </script>
 
 <template>
   <div class="app">
-    <p>
-      站點名稱搜尋: <input type="text" class="border" v-model="searchText">
-    </p>
+    <search :searchText="searchText" @change="updateSearchText" />
 
-    <table class="table table-striped">
-      <thead>
-        <tr>
-          <th @click="setSort('sno')">
-            #
-            <span v-show="currentSort === 'sno'">
-              <i class="fa" :class="isSortDesc ? 'fa-sort-desc' : 'fa-sort-asc'" aria-hidden="true"></i>
-            </span>
-          </th>
-          <th>
-            場站名稱
-          </th>
-          <th>
-            場站區域
-          </th>
-          <th @click="setSort('sbi')" class="pointer">
-            目前可用車輛
-            <span v-show="currentSort === 'sbi'">
-              <i class="fa" :class="isSortDesc ? 'fa-sort-desc' : 'fa-sort-asc'" aria-hidden="true"></i>
-            </span>
-          </th>
-          <th @click="setSort('tot')" class="pointer">
-            總停車格
-            <span v-show="currentSort === 'tot'">
-              <i class="fa" :class="isSortDesc ? 'fa-sort-desc' : 'fa-sort-asc'" aria-hidden="true"></i>
-            </span>
-          </th>
-          <th>
-            資料更新時間
-          </th>
-        </tr>
-      </thead>
-      <tbody>
-        <!-- 替換成 slicedUbikeStops -->
-        <tr v-for="s in slicedUbikeStops" :key="s.sno">
-          <td>{{ s.sno }}</td>
-          <!-- <td>{{ s.sna }}</td> -->
-          <td v-html="keywordsHighlight(s.sna, searchText)"></td>
-          <td>{{ s.sarea }}</td>
-          <td>{{ s.sbi }}</td>
-          <td>{{ s.tot }}</td>
-          <td>{{ timeFormat(s.mday) }}</td>
-        </tr>
-      </tbody>
-    </table>
+    <uBikeTable :slicedUbikeStops="slicedUbikeStops" :searchText="searchText" :currentSort="currentSort"
+      :isSortDesc="isSortDesc" @updateCurrentSort="updateCurrentSort" @updateIsSortDesc="updateIsSortDesc" />
   </div>
 
   <!-- 頁籤 -->
-  <nav v-if="pagerEnd > 0">
-    <ul class="pagination">
-      <li @click.prevent="setPage(currentPage - 1)" class="page-item">
-        <a class="page-link" href>Previous</a>
-      </li>
-
-      <li v-for="i in pagerEnd" :class="{ active: i + pagerAddAmount === currentPage }" :key="i"
-        @click.prevent="setPage(i + pagerAddAmount)" class="page-item">
-        <a class="page-link" href>{{ i + pagerAddAmount }}</a>
-      </li>
-
-      <li @click.prevent="setPage(currentPage + 1)" class="page-item">
-        <a class="page-link" href>Next</a>
-      </li>
-    </ul>
-  </nav>
+  <pagination :currentPage="currentPage" :totalPageCount="totalPageCount" :PAGINATION_MAX="PAGINATION_MAX"
+    @updateCurrentPage="updateCurrentPage" />
 </template>
 
 <style lang="scss" scoped>
 .app {
   padding: 1rem;
 }
+
 .pointer {
   cursor: pointer;
 }
+
 .pagination {
   display: flex;
   justify-content: center;
 }
+
 @media (max-width: 768px) {
   .sno {
-    max-width: 50px; word-wrap: break-word;
+    max-width: 50px;
+    word-wrap: break-word;
   }
-  .table td, .table th {
+
+  .table td,
+  .table th {
     padding: .5rem .25rem;
   }
 }
